backend/requirements.txt
fastapi
uvicorn[standard]
pillow
ultralytics
torch
torchvision
numpy
opencv-python
python-dotenv
firebase-admin
python-multipart==0.0.6
